Return the row count as an integer, not as text
count() is bigint, and the driver hands bigint to the workflow as a string.
Every other count in this router already goes through Number() at the point it
is read, which is how the gate has always survived it; rows_updated was new and
had no such wrapper, so a later strict comparison against 0 would have been
false for "0" and quietly done nothing.

Casting once at the source is cheaper than remembering to wrap it at every
future read.

+++ db/00-intake-router/say-the-write-was-refused.sql
@@ -13,4 +13,4 @@ SELECT
   $1::text                        AS gmail_message_id,
   $2::text                        AS _error,
   'Save triage'::text             AS _node,
-  (SELECT count(*) FROM updated)  AS rows_updated;
+  (SELECT count(*) FROM updated)::int AS rows_updated;
